requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-dotenv
python-multipart
google-adk
google-cloud-vision
sqlalchemy
aiosqlite
asyncpg
pytest
pytest-asyncio
sqlmodel
pandas
openpyxl
psycopg2-binary
